A tiny, and very naive, initial implementation.
In this implementation, we break the source text up into 'words' and
'whitespace', where a word is a maximal sequence of non-space
characters, and a whitespace is a maximal sequence of space
characters.  A 'space' character is one for which Data.Char.isSpace is
'True'.

We then use the Data.Algorithms.Diff.getDiff (from the Diff package)
to compute a difference that treats all whitespace sequences as
equal. We then produce the output by iterating over the diff as
follows

 - discard words unique to the first stream
 - include whitespace unique to the first stream

 - include words unique to the second stream
 - include whitespace unique to the second stream (it's
   probably separating new words)

 - include words common to both streams
 - include the actual whitespace from the first stream
   for whitespace that is common to both streams

On small examples, or when comparing files that haven't changed, this
is very fast. However, it falls off a cliff quite easily. I compared a
large Java source file with the result of pretty-printing it with Txl
(using the Txl Java 1.5 grammar). This strips out comments, including
a header comment at the top of the file, so the diff algorithm sees
differences immediately at the beginning of the streams. The two
sequences being compared are tens of thousands of 'word' and
'whitespace' sequences, so the diff ends up consuming hundreds of
megabytes of memory and takes an extremely long time -- in fact, I
couldn't wait and killed the process.

So, obviously we're going to have to do some things differently. For
one, we should recognize some of the constructs of a programming
language, like comments, character strings, separators, operators, and
various kinds of braces. In particular, we will need to treat comments
as a separate category all their own, since in the typical use case
the comments will be stripped from the 'new' file and must be restored
from the old one. We might also need a way to strip all comments and
white space from both streams, compare just the 'words'. For words
that appear in both streams, we could look up the position in the
first stream, find the adjacent text and comments, and insert them.

Second, we may end up having to abandon a diff algorithm based on
computing longest common sub-sequences. We might need to pre-process
with something like the BLAST algorithm in order to find long
identical sequences in the file, and do a 'real' diff just in the
places between those sequences.