Rework MinWSDiff.minWSDiff.
Rework 'minWSDiff' by

- diffing lists of word texts, instead of WordOrWS

- reinserting non-words, giving preference to the non-words
  from the old version, except for trailing non-words

At this point, the unit tests pass. The exception of preferring
the trailing non-words from the new version permits one
of the unit tests (diff of an empty sequence with X is X) to
pass. I may change the test, but not as part of this commit.

At this point, minwsd discards comments that appear only in
the original version, since comments appear to be sequences
of words unique to the original, and so are discarded. In
subsequent commits, I'll make comments into a special type
of non-word token, and require recognizing them in the
parser, perhaps with a language-specific option.

Unfortunately, the diff for this commit is rather nasty
to read; I ended up changing 'precedingNonWords' into
'nonWordGroups' and dropping the word positions. I also
deleted the now-unused 'extractLeadingWS' function, and
removing the now-unused EqWS newtype.