Add a 'Comment' variant to the 'Token' type.
We'll need to treat comments as a special kind of non-word.
This just introduces the extra variant, but nothing else.
In particular, the Arbitrary instance declaration for Tokens
does not generate Comments, since the parser and unparser
need to understand them.

Since different programming languages have different
comment syntax, we will need to have families of parser/unparser
pairs that understand the same comment syntax. Further, the
generators in the unit tests need to be language-specific too.

I believe I'll need to rework the code a little so that MinWSDiff
is parametrized by the parser and unparser functions, rather than
assuming a single parser and unparser module.

The existing parser and unparser are useful as a 'plain text'
case. Before introducing a parser and unparser for, say, C-style
programming languages, I'll do the necessary rework.

At this point, the unit tests still pass, but only because the
'arbitrary' generator doesn't generate Comments.