Add the 'Language' module, and use it to abstract away the parser and unparser.
The 'Language' module defines the 'Language' type, which holds a
parser and an unparser. The unparser probably never needs to be
changed in practice, but the parser can be tailored for different
programming languages. The 'defaultLanguage' uses the 'plain text'
parser that never produces comments.

In 'MinWSDiff', we now pass a 'Language' value to the main functions,
and no longer rely on the 'Parse' and 'Unparse' modules directly.
This allows client code to select a 'Language' that matches the input
or some user request.

Currently, we support only the 'plain text' language represented by
'defaultLanguage'. Shortly, I'll add support for Java, or perhaps more
generally for C-like languages.

Eventually, I believe 'Language' will need more features. For example,
since it's possible that a source-to-source transformer will move
existing code inside some construct that requires further indentation,
or outside such a construct, simply restoring the original non-words
would be wrong. We need to restore the original non-words *and*
correct the indentation.  However, to do that we'd need to detect when
indentation levels increase and decrease. (Note that for python
language support, this is essential, since getting the indentation
wrong breaks the code.) 'Language' will eventually need to provide a
way to do this kind of indentation correction, or a way to help do it.

For now, I'm just interested in getting back the original code when
the only change is to strip out comments and change the white space,
so the current definition should suffice.