Tests: prepare for more than one 'Langauge'.
The 'Tests' module was biased toward the parser for plain text.
In particular,

- the Arbitrary instance generated tokens that could be
  produced by the plain text parser

- the 'prop_unparseParse' property really only applies to
  plain text tokens

- the 'minimize' function also only applies to plain text
  tokens

I've turned the old 'arbitrary' generator into 'plainTextTokens',
and redefined 'arbitrary' so that the text in Words, WS, and
Comments is completely unrestricted. The correct operation of
'minWSDiff' doesn't depend on the particulars of the contents
of tokens, the new 'arbitrary' generator still works with the
properties

   prop_minWSDiffXXIsX
   prop_minWSDiffEmptyXIsX
   prop_minWSDiffContainsJustNewWords

Property 'prop_unparseParse' is now 'prop_plainTextUnparseParse',
and uses the 'defaultLanguage' definition rather than directly
importing the Parse and Unparse modules.

Finally, 'minimize' is now 'minimizePlainText', since it assumes
the token contents are consistent with a plain text parse.

After these changes, the tests still pass.