Remove the 'Parse' and 'Unparse' modules.
The 'Language' module is the only one that depends on the
'Parse' and 'Unparse' modules. Moreover, the 'Unparse' type
class defined in 'Unparse' is unnecessary. I've moved the
code from 'Parse' and 'Unparse' into language, and removed
the original modules. The 'Unparse' type class is now
gone, and 'defaultUnparse' has the obvious one-line definition.