Give 'Tokens' some convenience functions.
Other modules have to use Text.unpack and Text.pack a lot, but
that could be reduced by putting a few convenience functions
in the 'Tokens' module. I've added

  word    :: String -> Token
  ws      :: String -> Token
  comment :: String -> Token

and

  contentString :: Token -> String

and used these in the 'Language' and 'Test' modules. 'Language'
no longer needs to import Data.Text. 'Test' still does, but with
an additional refactoring in the 'minimizePlainText' function,
I should be able to eliminate the import there too.