Add a token parser for C-like languages.
This adds module C.hs, which implements a 'Language' for C-like
languages, and uses it as the default instead of the 'plain text'
language. With this language definition I can restore the white space
and comments on a sample Java file (about 4250 lines) in a little
under six seconds. That's not terribly fast, but it's a start.

The parser in C.hs uses a combination of hand-rolled parsing and
regular expressions from regex-tdfa.  Some preliminary profiling seems
to show that the regular expressions really slow things down. In my
first crack at the 'afterDot' function, I was using a regular
expression to capture the possible fractional number following the
dot. If the expression failed to match, the dot is a word by itself;
otherwise, it starts a fractional number. It turns out that failing to
match the regular expression is much, much slower than checking for a
digit and only then using a regex to find the rest of the number. This
optimization has a huge performance impact, because the dot is used
for field accesses and method calls, and these uses are far more
frequent than appearances in floating-point literals. Before doing it,
it took almost seven seconds to process a 990-line file! Some
profiling showed that a lot of the time was being spent in the regex
matching.

To really improve the parser's performance, I may have to go
completely hand-rolled. That's too bad, because using regex matching
could give an easy way for a user to specify the lexical structure of
an unsupported language using a file of regular expressions. But for
common languages, perhaps hand-rolled is the way to go anyway.