Switch to using regex-pcre from regex-tdfa.
In the C-like langauge parser, the regular expressions I've
used should yield equivalent results when interpreted as
Posix or Perl-style regular expressions. regex-tdfa is
rather slow, so I decided to try changing to regex-pcre.

Without any other changes, I can restore the comments
and whitespace in a 4250-line original file in about a second,
according to 'time':

real	0m0.981s
user	0m0.940s
sys	0m0.030s

Using regex-tdfa in the last commit, doing the same thing took
over six seconds:

real	0m6.367s
user	0m6.230s
sys	0m0.100s

So switching definitely seems like a win. It's probably worthwhile
to experiment with a hand-written lexer to see how it compares.

According to

  http://www.haskell.org/haskellwiki/Regular_expressions

some versions of GHC have a bug that causes the performance of
regex-tdfa to tank; apparently, on each match the regular
expression is recompiled. The performance here is bad enough
that I wonder if this is the problem. According to the
benchmarks on that page, regex-tdfa shouldn't be as fast as
regex-pcre, but it shouldn't be so outclassed.