C.hs: restructure to permit arbtirary comment parsers
Lots of languages use C-like strings, numeric literals and
identifiers; but they're most likely to vary in the comment syntax.

We now generalize the `cParser` to `cLikeParser`, which takes a
comment parser with type

```
String -> Maybe (Token, String)
```

as an argument. The `defaultCommentParser` uses the defaults.

In the process, we refactor the existing token parsers so they also
have the type

```
String -> Maybe (Token, String)
```

and are not themselves responsible for parsing the rest of the input.