minwsd.hs: add support for the 'puppet' language
The 'puppet' language is for the Puppet infrastructure configuration
language, and is essentially C-like but allows both C-style and
shell-style commments.

We add support for it directly in the main executable, though we
should probably move the set of supported languages into its own
module. I'll leave that for a future commit.