Add some parse-unparse tests, and fix a bug in 'C.hs'
For any language definition `lang`, we should have

```
(unparser lang) . (parser lang)
```

be the identity function on strings. This is easy enough to test, so
I've added tests for `plainText`, `cLanguage`, and `shellLanguage`; on
the first run, this found a bug in `cLanguage` and `shellLanguage`
with the example input `"1\n1"`.

The problem turns out to be the regular expressions used in `C.hs`, in
particular the handling of a leading `^` in the expressions. By
default, `^` will match immediately after any newline, not necessarily
at the start of the input string. Here's how the falsifiable example
works.

The first character `1` is a digit, so we end up calling

```
afterDigit '1' "\n1"
```

Because we used `makeRegex`, which uses the default compilation
options, the regular expression

```
^([0-9]+(\\.[0-9]*)?|\\.[0-9]*)([Ee][-+]?[0-9]+)?
```

matches the `1` immediately after the newline, since the `^` matches
right after either the beginning of the string or any newline. We skip
over the newline character, and construct a token from the text "11".

After a little reading in the Haddock documentation for
`Text.Regex.Base.RegexLike`, `Text.Regex.PCRE.Wrap`, and the PCRE
library web pages, it looks like we need to use the compilation option
`compFirstLine`. I've created a function `makeTokenRegex` that uses
`makeRegexOpts` with `compFirstLine` and the default execution
options. Now the tests seem to pass -- at least after several runs --
and I've verified in GHCI that the input "1\n1" now reads as three
tokens and unparses to "1\n1" again.

Note that this error has existed here since 2011, and I found it
immediately after writing a trivial test. I've used it a bit on real
source files without trouble, but just because I was lucky.

I've also found with a few trials on a test input that the program
speeds up a small fraction of a second. I'll need to do some more
systematic tests to verify that this isn't just luck.